Fix timeouts while unpacking data from HTTP streams
The idea is to open each .app file one after the other, and read the files contained inside in sequential order.
Previously, all .app streams would be kept open in parallel (even if all contents were already read from one of the files) until the entire unpacking process was done,
which resulted in timeouts for files that were not read for a longer period of time.